requirements.txt
boto3==latest
botocore==latest
deep-translator==latest
PyGithub==latest